P4-3392 initial steps moving towards using BigDecimal instead of double for price adjustments. More to do at time of commit.

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/domain/auditing/AuditableEvent.kt
@@ -123,7 +123,7 @@ data class AuditableEvent(
     ) = AuditableEvent(
       type = AuditEventType.JOURNEY_PRICE_BULK_ADJUSTMENT,
       username = authentication?.name ?: terminal,
-      metadata = AnnualPriceAdjustmentMetadata(supplier, effectiveYear, multiplier, details)
+      metadata = AnnualPriceAdjustmentMetadata(supplier, effectiveYear, multiplier.toBigDecimal(), details)
     )
 
     fun mapLocation(location: Location) =
